+++ src/nd.hpp
@@ -119,6 +119,7 @@ public:
 	static Ref<NDArray> squeeze(const Variant& v, const Variant& axes);
 	static Ref<NDArray> expand_dims(const Variant& v, int64_t axis);
 	static Ref<NDArray> roll(const Variant& v, const Variant& shift, const Variant& axis);
+	static Ref<NDArray> repeat(const Variant& v, const Variant& repeats, const Variant& axis);
 
 	// Complex.
 	static Ref<NDArray> real(const Variant& a);